requirements.txt
python-dotenv>=0.20.0
selenium>=4.3.0
